base models projects/glass-box-portal/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
playwright
pillow